+++ bindings/dart/lib/src/doc/document.dart
@@ -1,3 +1,6 @@
+import 'package:xayn_ai_ffi_dart/src/ffi/genesis.dart'
+    show CFeedback, CRelevance;
+
 /// A document relevance level.
 enum Relevance {
   low,
@@ -5,11 +8,71 @@ enum Relevance {
   high,
 }
 
+extension RelevanceInt on Relevance {
+  /// Gets the discriminant.
+  int toInt() {
+    switch (this) {
+      case Relevance.low:
+        return CRelevance.Low;
+      case Relevance.medium:
+        return CRelevance.Medium;
+      case Relevance.high:
+        return CRelevance.High;
+      default:
+        throw UnsupportedError('Undefined enum variant.');
+    }
+  }
+
+  /// Creates the relevance level from a discriminant.
+  static Relevance fromInt(int idx) {
+    switch (idx) {
+      case CRelevance.Low:
+        return Relevance.low;
+      case CRelevance.Medium:
+        return Relevance.medium;
+      case CRelevance.High:
+        return Relevance.high;
+      default:
+        throw UnsupportedError('Undefined enum variant.');
+    }
+  }
+}
+
 /// A user feedback level.
 enum Feedback {
   relevant,
   irrelevant,
   none,
+}
+
+extension FeedbackInt on Feedback {
+  /// Gets the discriminant.
+  int toInt() {
+    switch (this) {
+      case Feedback.relevant:
+        return CFeedback.Relevant;
+      case Feedback.irrelevant:
+        return CFeedback.Irrelevant;
+      case Feedback.none:
+        return CFeedback.None;
+      default:
+        throw UnsupportedError('Undefined enum variant.');
+    }
+  }
+
+  /// Creates the feedback level from a discriminant.
+  static Feedback fromInt(int idx) {
+    switch (idx) {
+      case CFeedback.Relevant:
+        return Feedback.relevant;
+      case CFeedback.Irrelevant:
+        return Feedback.irrelevant;
+      case CFeedback.None:
+        return Feedback.none;
+      default:
+        throw UnsupportedError('Undefined enum variant.');
+    }
+  }
 }
 
 /// The document history.
